# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce API",
    "version": "1.0.0",
    "description": "API documentation for the E-Commerce platform"
  },
  "servers": [
    {
      "url": "http://localhost:3300"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/general-settings": {
      "get": {
        "summary": "Get general settings",
        "tags": [
          "GeneralSettings"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "General settings data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "type": "object",
                      "description": "The settings data"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update general settings",
        "tags": [
          "GeneralSettings"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "link_instagram": "https://instagram.com/yourstore",
                  "link_whatsapp": "https://wa.me/628123456789",
                  "alamat": "Jl. Contoh No. 123",
                  "email": "store@example.com",
                  "service_fee_mode": "buyer",
                  "waktu_operasi_toko": "08:00 - 17:00"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated or created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ld/home": {
      "get": {
        "summary": "Get Little dimple",
        "tags": [
          "Little Dimple"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Little dimple data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "type": "object",
                      "description": "The settings data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/ld/contact-us": {
      "post": {
        "summary": "Submit a contact us message",
        "tags": [
          "Little Dimple"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "subject",
                  "message"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "subject": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact message submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contact": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "subject": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/ld/warranty": {
      "post": {
        "summary": "Check if warranty exists by code",
        "tags": [
          "Warranty"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kode_warranty": {
                    "type": "string"
                  }
                },
                "required": [
                  "kode_warranty"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Warranty found or not found"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/ld/manual-warranty": {
      "post": {
        "summary": "Insert manual warranty data",
        "tags": [
          "Warranty"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "nohp": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "province": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "district": {
                    "type": "string"
                  },
                  "subdistrict": {
                    "type": "string"
                  },
                  "postalcode": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "attachment_receipt": {
                    "type": "string",
                    "format": "binary"
                  },
                  "attachment_barcode": {
                    "type": "string",
                    "format": "binary"
                  },
                  "attachment_product": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Manual warranty submitted successfully"
          }
        }
      }
    },
...
  },
...
}